program ADC h/w directly

+++ firmware/src/adc.h
@@ -2,4 +2,4 @@
 #include <stdint.h>
 
 void adc_init();
-uint16_t adc_read_mV();
+uint32_t adc_read_mV();
